# Update xxhash to v2.1.1 and improve pooling