I started on this because I didn't like to see my program linking two versions of xxhash.
The way it's used in Alertmanager it makes no practical difference; I just wanted to tidy up.

Then, while I was looking at it, I decided the buffer pooling could be made more idiomatic.
By putting pointer-valued objects into the pool we avoid boxing a slice value; this saves one memory allocation every time.

I added a benchmark to illustrate:
```
name         old time/op    new time/op    delta
HashAlert-4     529ns ± 1%     490ns ± 2%   -7.29%  (p=0.008 n=5+5)

name         old alloc/op   new alloc/op   delta
HashAlert-4     96.0B ± 0%     72.0B ± 0%  -25.00%  (p=0.008 n=5+5)

name         old allocs/op  new allocs/op  delta
HashAlert-4      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.008 n=5+5)
```

(I also tried re-coding using `xxhash.New()`, but that is slower, probably because it uses the Go implementation and not the assembly one).